progress

+++ func_impl_tr_hub.c
@@ -28,34 +28,62 @@ void send_sig(){
         kill(monitor_pid,SIGUSR1);
 }
 
+void handle_sigchld(int sig)
+{
+    if(monitor_stopping)
+    {
+        monitor_pid=-1;
+        monitor_stopping=0;
+        printf("Monitor stopped\n");
+    }
+}
+
 
 void stop_monitor() {
+    struct sigaction sigchld;
+    sigchld.sa_handler = handle_sigchld;
+    sigchld.sa_flags=0;
+    sigemptyset(&sigchld.sa_mask);
+    if((sigaction(SIGCHLD, &sigchld, NULL))==-1)
+    {
+        perror("Error sigaction");
+        exit(EXIT_FAILURE);   
+    }
+
     if (monitor_running == 0) {
         printf("ERR: Monitor is not running\n");
         return;
     }
-    printf("Monitor will be stopped...\n");
-    usleep(5000000);
-    kill(monitor_pid, SIGTERM);
     monitor_stopping = 1;
-    
-    while( monitor_pid != -1 ){
-        char cmd[31];
-        if(scanf("%s", cmd) == 1){
-            printf("Command <%s> couldn't be processed while monitor is stopping\n", cmd);
+    printf("Monitor will be stopped...\n");
+    int pid = fork();
+    if( pid == 0 ){
+        usleep(5000000);
+        kill(monitor_pid, SIGTERM);
+        exit(0);
+    }
+    else if(pid > 0){
+        while( monitor_pid != -1 ){
+            char cmd[31];
+            if(scanf("%s", cmd) == 1){
+                printf("Command <%s> couldn't be processed while monitor is stopping\n", cmd);
+            }
+        }
+
+        
+
+        int status;
+        waitpid(monitor_pid,&status, 0);
+
+        if (WIFEXITED(status)) {
+            printf("Monitor exited with status %d\n", WEXITSTATUS(status));
+        } else if (WIFSIGNALED(status)) {
+            printf("Monitor terminated by signal %d\n", WTERMSIG(status));
+        } else {
+            printf("Monitor exited abnormally\n");
         }
     }
-
-    int status;
-    waitpid(monitor_pid,&status, 0);
-
-    if (WIFEXITED(status)) {
-        printf("Monitor exited with status %d\n", WEXITSTATUS(status));
-    } else if (WIFSIGNALED(status)) {
-        printf("Monitor terminated by signal %d\n", WTERMSIG(status));
-    } else {
-        printf("Monitor exited abnormally\n");
-    }
+    
 }
 
 void handle_sigusr1(int sig){
@@ -64,6 +92,7 @@ void handle_sigusr1(int sig){
         perror("error opening file");
         exit(EXIT_FAILURE);
     }
+    int calc_score = 0;
     char args[31];
     char *v_arg[5];
     int i = 0;
@@ -75,6 +104,10 @@ void handle_sigusr1(int sig){
 
     strcpy(v_arg[i++], "treasure_manager");
     while(fgets(args,sizeof(args),f) != NULL){
+        if(strcmp(args,"calc") == 0){
+            calc_score =1;
+            break;
+        }
         args[strcspn(args, "\n")] = 0;
         v_arg[i] = malloc(strlen(args) + 1);
         if( v_arg[i] == NULL ){
@@ -90,43 +123,112 @@ void handle_sigusr1(int sig){
 
     fclose(f); 
 
-    int pid = vfork();
-    if(pid == 0 ){
-        char cwd[1024];
-        getcwd(cwd,sizeof(cwd));
+    int pfd[2];
+    if(pipe(pfd) < 0 ){
+        perror("Pipe creation failed\n");
+        exit(EXIT_FAILURE);
+    }
 
-        char exec_path[1024];
-        int written = snprintf(exec_path, sizeof(exec_path), "%s/treasure_manager",cwd);
-        if (written < 0 || written >= sizeof(exec_path)) {
-            perror("Path too long\n");
+    if(!calc_score){
+
+        int pid = fork();
+        if(pid == 0 ){
+
+            close(pfd[0]);
+            if(dup2(pfd[1], 1) < 0){
+                perror("dup2 error\n");
+                exit(EXIT_FAILURE);
+            }
+
+            char cwd[1024];
+            getcwd(cwd,sizeof(cwd));
+
+            char exec_path[1024];
+            int written = snprintf(exec_path, sizeof(exec_path), "%s/treasure_manager",cwd);
+            if (written < 0 || written >= sizeof(exec_path)) {
+                perror("Path too long\n");
+                exit(EXIT_FAILURE);
+            }
+
+            execvp(exec_path,v_arg);
+            perror("exec failed");
+            exit(EXIT_FAILURE);
+        }
+        else if(pid > 0 ){
+            FILE *stream;
+            stream = fdopen(pfd[0], "r");
+            close(pfd[1]);
+            if( stream == NULL ){
+                perror("fdopen error\n");
+                exit(EXIT_FAILURE);
+            }
+            char string[1024];
+            // if(unlink("out.txt") == -1 && errno != ENOENT){
+            //     perror("Failed to delete file");
+            //     exit(EXIT_FAILURE);
+            // }
+            // FILE *out = fopen("out.txt", "a");
+            // if(out == NULL ){
+            //     perror("Error when opening file\n");
+            //     exit(EXIT_FAILURE);
+            // }
+            printf("Output din proces parinte cu pipe:\n");
+            while( fgets(string,sizeof(string), stream) != NULL){
+                printf("%s", string);
+            }
+            close(pfd[0]);
+            int status;
+            waitpid(pid,&status, 0);  
+        }
+        else{
+            perror("fork failed");
             exit(EXIT_FAILURE);
         }
 
-        execvp(exec_path,v_arg);
-        perror("exec failed");
-        exit(EXIT_FAILURE);
-    }
-    else if(pid > 0 ){
-        int status;
-        waitpid(pid,&status, 0);
+        for (int i = 0; v_arg[i] != NULL; i++) {
+            free(v_arg[i]);
+        }
     }
     else{
-        perror("fork failed");
-        exit(EXIT_FAILURE);
-    }
-    for (int i = 0; v_arg[i] != NULL; i++) {
-        free(v_arg[i]);
+        int pid = fork();
+        if(pid == 0 ){
+
+            close(pfd[0]);
+            if(dup2(pfd[1], 1) < 0){
+                perror("dup2 error\n");
+                exit(EXIT_FAILURE);
+            }
+
+            char cwd[1024];
+            getcwd(cwd,sizeof(cwd));
+
+            char exec_path[1024];
+            int written = snprintf(exec_path, sizeof(exec_path), "%s/calc_score",cwd);
+            if (written < 0 || written >= sizeof(exec_path)) {
+                perror("Path too long\n");
+                exit(EXIT_FAILURE);
+            }
+            execvp(exec_path,v_arg);
+            perror("exec failed");
+            exit(EXIT_FAILURE);
+        }
+        else if(pid> 0){
+            
+        }
+        else{
+
+        }
     }
 
 }
 
 void monitor_proc(){
     struct sigaction sa_sigusr1;
-    sa_sigusr1.sa_handler = handle_sigusr1; //seteaza campul din structul sigaction cu un handler
-    sigemptyset(&sa_sigusr1.sa_mask); //seteaza masca pe 0 pt a nu bloca semnale neintentionat
+    sa_sigusr1.sa_handler = handle_sigusr1;
+    sigemptyset(&sa_sigusr1.sa_mask); 
     sa_sigusr1.sa_flags = 0; 
 
-    sigaction(SIGUSR1, &sa_sigusr1, NULL); //sigaction in sine, ii spuen procesului ce sa faca daca primeste semnalul SIGUSR1 prin variabila sa_sigusr1 si campurile ei
+    sigaction(SIGUSR1, &sa_sigusr1, NULL);
 
     while(1){
         pause();
@@ -153,13 +255,3 @@ void start_monitor(){
 }
 
 
-void handle_sigchld(int sig)
-{
-    if(monitor_stopping)
-    {
-        monitor_pid=-1;
-        // printf("Monitor stopped\n");
-
-        monitor_stopping=0;
-    }
-}
